build: improve manifest script readability
