+++ scripts/build-manifest/fix-inherited-field-types.js
@@ -12,22 +12,7 @@ const INHERITED_FIELDS = [
 const types = new Map();
 
 export default () => ({
-  name: 'INTL-ELEMENTS: Fix inherited field types and aliased types',
-
-  analyzePhase: ({ts, node}) => {
-    if (node.kind === ts.SyntaxKind.TypeAliasDeclaration) {
-      const {fileName} = node.parent;
-      const ref = {
-        name: node.name.getText(),
-        type: node.type.getText(),
-      };
-      if (types.has(fileName)) {
-        types.get(fileName).push(ref);
-      } else {
-        types.set(fileName, [ref]);
-      }
-    }
-  },
+  name: 'INTL-ELEMENTS: Fix inherited field types',
 
   moduleLinkPhase: ({moduleDoc}) => {
     moduleDoc.declarations.forEach(declaration => {
@@ -42,33 +27,10 @@ export default () => ({
   packageLinkPhase: ({customElementsManifest: manifest}) => {
     manifest.modules.forEach(module => {
       module.declarations.forEach(declaration => {
-        if (!declaration.members) {
-          return;
-        }
-
-        declaration.members.forEach(member => {
+        declaration.members?.forEach(member => {
           if (member.originalType) {
             member.type = Object.assign({}, member.originalType);
             delete member.originalType;
-          }
-          if (types.has(module.path)) {
-            let ref;
-            switch (member.kind) {
-              case 'method':
-                ref = types.get(module.path)
-                    .find(type => type.name === member.return?.type?.text);
-                if (ref) {
-                  member.return.type.text = ref.type;
-                }
-                break;
-              case 'field':
-                ref = types.get(module.path)
-                    .find(type => type.name === member.type?.text);
-                if (ref) {
-                  member.type.text = ref.type;
-                }
-                break;
-            }
           }
         });
       });
