+++ scripts/build-manifest/index.js
@@ -1,15 +1,18 @@
 import customJsDocTags from './custom-jsdoc-tags.js';
+import fixAliasedMemberTypes from './fix-aliased-member-types.js';
 import fixInheritedFieldTypes from './fix-inherited-field-types.js';
 import optionPropertyDecorator from './option-property-decorator.js';
 import removeMembers from './remove-members.js';
 import removeInternalClasses from './remove-internal-classes.js'
-import sort from './sort.js';
+import sortManifest from './sort-manifest.js';
 
+// The order of these plugins matters
 export default [
   customJsDocTags(),
   optionPropertyDecorator(),
   fixInheritedFieldTypes(),
+  fixAliasedMemberTypes(),
   removeMembers(),
   removeInternalClasses(),
-  sort(),
+  sortManifest(),
 ];
