+++ scripts/build-manifest/sort-manifest.js
